feat: tableau traffic to views asset
