+++ src/teamster/libraries/couchdrop/sensors.py
@@ -35,7 +35,7 @@ def build_couchdrop_sftp_sensor(
     if exclude_dirs is None:
         exclude_dirs = []
 
-    base_job_name = f"{code_location}_couchdrop_sftp_asset_job"
+    base_job_name = f"{code_location}__couchdrop__sftp_asset_job"
 
     keys_by_partitions_def = defaultdict(set[AssetKey])
