+++ tests/assets/sftp/test_assets_sftp_couchdrop.py
@@ -1,11 +1,17 @@
 import random
 
-from dagster import materialize
+from dagster import (
+    DynamicPartitionsDefinition,
+    MultiPartitionsDefinition,
+    _check,
+    instance_for_test,
+    materialize,
+)
 
 from teamster.core.resources import SSH_COUCHDROP, get_io_manager_gcs_avro
 
 
-def _test_asset(asset, partition_key=None):
+def _test_asset(asset, partition_key=None, instance=None):
     if partition_key is not None:
         pass
     elif asset.partitions_def is not None:
@@ -34,6 +40,76 @@ def _test_asset(asset, partition_key=None):
 
     assert extras is not None
     assert extras.text == ""
+
+
+def test_performance_management_observation_details_kipptaf():
+    from teamster.code_locations.kipptaf.performance_management.assets import (
+        observation_details,
+    )
+
+    _test_asset(asset=observation_details)
+
+
+def test_adp_payroll_general_ledger_file_kipptaf():
+    from teamster.code_locations.kipptaf.adp.payroll.assets import general_ledger_file
+
+    date_key = "20241130"
+    group_code_key = "47S"
+
+    partitions_def = _check.inst(
+        obj=general_ledger_file.partitions_def, ttype=MultiPartitionsDefinition
+    )
+
+    date_partitions_def = _check.inst(
+        obj=partitions_def.get_partitions_def_for_dimension("date"),
+        ttype=DynamicPartitionsDefinition,
+    )
+
+    partitions_def_name = _check.not_none(value=date_partitions_def.name)
+
+    with instance_for_test() as instance:
+        instance.add_dynamic_partitions(
+            partitions_def_name=partitions_def_name, partition_keys=[date_key]
+        )
+
+        _test_asset(
+            asset=general_ledger_file,
+            instance=instance,
+            partition_key=f"{date_key}|{group_code_key}",
+        )
+
+
+def test_fldoe_fast_kippmiami():
+    from teamster.code_locations.kippmiami.fldoe.assets import fast
+
+    _test_asset(asset=fast)
+
+
+def test_fldoe_fsa_kippmiami():
+    from teamster.code_locations.kippmiami.fldoe.assets import fsa
+
+    _test_asset(asset=fsa)
+
+
+def test_fldoe_eoc_kippmiami():
+    from teamster.code_locations.kippmiami.fldoe.assets import eoc
+
+    _test_asset(asset=eoc)
+
+
+def test_fldoe_science_kippmiami():
+    from teamster.code_locations.kippmiami.fldoe.assets import science
+
+    _test_asset(asset=science)
+
+
+def test_fldoe_fte_kippmiami():
+    from teamster.code_locations.kippmiami.fldoe.assets import fte
+
+    _test_asset(
+        asset=fte,
+        # partition_key="25|2",
+    )
 
 
 def test_pearson_njgpa_kippcamden():
@@ -94,3 +170,9 @@ def test_pearson_student_list_report_kippnewark():
     from teamster.code_locations.kippnewark.pearson.assets import student_list_report
 
     _test_asset(asset=student_list_report)
+
+
+def test_tableau_traffic_to_views_kipptaf():
+    from teamster.code_locations.kipptaf.tableau.assets import traffic_to_views
+
+    _test_asset(asset=traffic_to_views)
